requirements.txt
flask
requests
pandas
cachetools
python-dotenv 